split off python dist generation test base into a util subdir
