+++ tests/python/pants_test/backend/python/tasks/util/BUILD
@@ -1,0 +1,16 @@
+python_library(
+  dependencies=[
+    '3rdparty/python:future',
+    '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/backend/native',
+    'src/python/pants/backend/native/targets',
+    'src/python/pants/backend/python:plugin',
+    'src/python/pants/backend/python/targets',
+    'src/python/pants/backend/python/tasks',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:process_handler',
+    'tests/python/pants_test:int-test',
+    'tests/python/pants_test/backend/python/tasks:python_task_test_base',
+    'tests/python/pants_test/engine:scheduler_test_base',
+  ]
+)
